fast-api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
langchain==0.3.22
langchain-openai==0.3.12
redis==5.2.1
psycopg2-binary==2.9.10
python-jose==3.4.0
pydantic==2.11.2
uuid==1.30